refactor: extract SentimentMath shared utility from duplicated handlers
DecayWeightedAverage, CalculateTrendDirection, and CalculateDispersion
were copy-pasted across trending and watchlist handlers. Extracted into
a single SentimentMath static class for reuse.
