+++ Application/Features/Sentiment/Queries/GetTrendingSymbols/GetTrendingSymbolsQueryHandler.cs
@@ -7,8 +7,6 @@ namespace Application.Features.Sentiment.Queries.GetTrendingSymbols;
 public class GetTrendingSymbolsQueryHandler(ISentimentRepository repository)
     : IRequestHandler<GetTrendingSymbolsQuery, IReadOnlyList<TrendingSymbolDto>>
 {
-    private const int DefaultHalfLifeHours = 72;
-
     public async Task<IReadOnlyList<TrendingSymbolDto>> Handle(
         GetTrendingSymbolsQuery query,
         CancellationToken ct)
@@ -66,8 +64,8 @@ public class GetTrendingSymbolsQueryHandler(ISentimentRepository repository)
         var current  = analyses.Where(a => a.AnalyzedAt >= midpoint).ToList();
         var previous = analyses.Where(a => a.AnalyzedAt <  midpoint).ToList();
 
-        var currentAvg  = DecayWeightedAverage(current, now);
-        var previousAvg = DecayWeightedAverage(previous, midpoint);
+        var currentAvg  = SentimentMath.DecayWeightedAverage(current, now);
+        var previousAvg = SentimentMath.DecayWeightedAverage(previous, midpoint);
 
         var delta = Math.Round(currentAvg - previousAvg, 4);
         currentAvg  = Math.Round(currentAvg,  4);
@@ -80,95 +78,10 @@ public class GetTrendingSymbolsQueryHandler(ISentimentRepository repository)
             _    => "flat"
         };
 
-        // Trend via linear regression over the full window
-        var trend = CalculateTrendDirection(analyses);
-
-        // Dispersion: weighted std dev over all analyses
-        var dispersion = CalculateDispersion(analyses, now);
+        var trend = SentimentMath.CalculateTrendDirection(analyses);
+        var dispersion = SentimentMath.CalculateDispersion(analyses, now);
 
         return new TrendingSymbolDto(symbol, currentAvg, previousAvg, delta, direction,
             trend, dispersion, analyses.Count);
-    }
-
-    private static string CalculateTrendDirection(IReadOnlyList<SentimentAnalysis> analyses)
-    {
-        if (analyses.Count < 2)
-            return "Stable";
-
-        var origin = analyses.Min(a => a.AnalyzedAt);
-        var points = analyses
-            .Select(a => ((a.AnalyzedAt - origin).TotalHours, a.Score.Value))
-            .ToList();
-
-        var n = points.Count;
-        var sumX = points.Sum(p => p.Item1);
-        var sumY = points.Sum(p => p.Item2);
-        var sumXY = points.Sum(p => p.Item1 * p.Item2);
-        var sumX2 = points.Sum(p => p.Item1 * p.Item1);
-
-        var denominator = n * sumX2 - sumX * sumX;
-        if (Math.Abs(denominator) < double.Epsilon)
-            return "Stable";
-
-        var slope = (n * sumXY - sumX * sumY) / denominator;
-
-        return slope switch
-        {
-            > 0.002  => "Improving",
-            < -0.002 => "Deteriorating",
-            _        => "Stable"
-        };
-    }
-
-    private static double CalculateDispersion(IReadOnlyList<SentimentAnalysis> analyses, DateTime now)
-    {
-        if (analyses.Count < 2)
-            return 0.0;
-
-        var totalWeight = 0.0;
-        var weightedSum = 0.0;
-
-        foreach (var a in analyses)
-        {
-            var ageHours = Math.Max(0, (now - a.AnalyzedAt).TotalHours);
-            var weight = Math.Exp(-Math.Log(2) / DefaultHalfLifeHours * ageHours);
-            totalWeight += weight;
-            weightedSum += weight * a.Score.Value;
-        }
-
-        if (totalWeight <= 0) return 0.0;
-        var mean = weightedSum / totalWeight;
-
-        var varianceSum = 0.0;
-        foreach (var a in analyses)
-        {
-            var ageHours = Math.Max(0, (now - a.AnalyzedAt).TotalHours);
-            var weight = Math.Exp(-Math.Log(2) / DefaultHalfLifeHours * ageHours);
-            varianceSum += weight * Math.Pow(a.Score.Value - mean, 2);
-        }
-
-        return Math.Round(Math.Sqrt(varianceSum / totalWeight), 4);
-    }
-
-    private static double DecayWeightedAverage(
-        IReadOnlyList<SentimentAnalysis> analyses,
-        DateTime referenceTime)
-    {
-        if (analyses.Count == 0)
-            return 0.0;
-
-        var totalWeight = 0.0;
-        var weightedSum = 0.0;
-
-        foreach (var a in analyses)
-        {
-            var ageHours = (referenceTime - a.AnalyzedAt).TotalHours;
-            if (ageHours < 0) ageHours = 0;
-            var weight = Math.Exp(-Math.Log(2) / DefaultHalfLifeHours * ageHours);
-            totalWeight += weight;
-            weightedSum += weight * a.Score.Value;
-        }
-
-        return totalWeight > 0 ? weightedSum / totalWeight : 0.0;
     }
 }
